% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/before_install.R
\name{before_install}
\alias{before_install}
\title{To run before newly installing inbobosvitaliteit package}
\usage{
before_install()
}
\value{
de nodige files worden in de package directory geschreven
}
\description{
To run before newly installing inbobosvitaliteit package
}
\examples{
\dontrun{
library(inbolims)
setwd(here::here()
before_install()
}
}
