% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bomen_calc_orig.R
\name{bomen_calc_orig}
\alias{bomen_calc_orig}
\title{Bereken de tabellen voor het meeste van de jaarlijkse en symptoomanalyses}
\usage{
bomen_calc_orig(
  x,
  group = c("Jaar"),
  group2 = NULL,
  respons = NULL,
  uniquecount = NULL,
  na.action = na.omit
)
}
\arguments{
\item{x}{dataset met minstens de variabele MetingKey en AantastingsKey waarop de berekeningen gebeuren.
MetingKey indentificeert unieke bomen per jaar, AantastingsKey identificeert aantastingen uniek, zodat aantastingen met meerdere specificaties of oorzaken niet dubbel geteld worden}

\item{group}{een vector of lijst met vectoren voor de primaire groepering als basis voor percentageberekenig}

\item{group2}{laat nog extra groepering toe, is meestal de laatste categorische variabele, waarover we de percentages willen weten, dus dit is meestal een variabele met x categorieën, en het percentage per categorie wordt berekend}

\item{respons}{indien ingevuld wordt voor deze variabele de mean, median, sd en se berekend}

\item{uniquecount}{indien ingevuld wordt per groep (group+group2) geteld hoeveel unieke waarden deze variabele heeft}

\item{na.action}{functie die aangeeft wat er moet gebeuren met de missings. na.omit is de logische default}
}
\value{
een tabel met per combinatie van groepen (in group meegegeven) een inschatting van het aantal bomen,
aantal records, unieke waarden en de overeenkomstige percentages, eventueel nog aangevuld met statistieken op een responsvariabele (mean, se, sd, mediaan)
}
\description{
Bereken de tabellen voor het meeste van de jaarlijkse en symptoomanalyses
}
