% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bosvitaliteit_connect.R
\name{bosvitaliteit_connect}
\alias{bosvitaliteit_connect}
\title{Connect to bosvitaliteit DB}
\usage{
bosvitaliteit_connect()
}
\value{
DBI DB connection object
}
\description{
Connect to bosvitaliteit DB
}
