% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sen_slope.R, R/pred_sen_slope.R
\name{calc_sen_slope}
\alias{calc_sen_slope}
\title{Calculate sen slope on tree data}
\usage{
calc_sen_slope(data, sen_boot = 0)

calc_sen_slope(data, sen_boot = 0)
}
\arguments{
\item{data}{the data to base the calculations on}

\item{sen_boot}{the amount of bootstrap samples for the confidence intervals. When 0 no bootstrap is performed and the values are returned as is}
}
\value{
data.frame with sen-slope results

data.frame with sen-slope results
}
\description{
Calculate sen slope on tree data

Calculate sen slope on tree data
}
