% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgmt_functions.R
\name{copy_scripts_and_sql_to}
\alias{copy_scripts_and_sql_to}
\title{Kopieer scripts, sql en csv data bestanden naar lokale directory}
\usage{
copy_scripts_and_sql_to(target)
}
\arguments{
\item{target}{directory waarin de files weggeschreven worden}
}
\value{
lokale kopie van scripts en sql en csv data in de scripts en data directory onder de target
}
\description{
Kopieer scripts, sql en csv data bestanden naar lokale directory
}
