% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_file_structure.R
\name{generate_file_structure}
\alias{generate_file_structure}
\title{Generate the necessary folders for the project exports}
\usage{
generate_file_structure(root = getwd())
}
\arguments{
\item{root}{directory where the file structure should start from}
}
\value{
directories created on the file system
}
\description{
Generate the necessary folders for the project exports
}
