% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_generator_functions.R
\name{generate_sql_files}
\alias{generate_sql_files}
\title{Generate the SQL files used to communicate with the database}
\usage{
generate_sql_files(target = "extdata")
}
\arguments{
\item{target}{target directory where the files should be placed}
}
\value{
sql files created in the target directory
}
\description{
Generate the SQL files used to communicate with the database
}
