% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_generator_functions.R
\name{generate_tree_species_data}
\alias{generate_tree_species_data}
\title{Generate tree species metadatafile for use in reporting}
\usage{
generate_tree_species_data(target = "extdata")
}
\arguments{
\item{target}{directory to export the species information to}
}
\value{
a csv (;) file generated with columns SPEC_EUR_CDE (European species code), SPEC_DES (scientific name), species, species_main_cat (report categorisation general), species_sub_cat (more detailed categorisation), species_name (English name), species_order (Reporting order)
}
\description{
Generate tree species metadatafile for use in reporting
}
