% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{get_script_location}
\alias{get_script_location}
\title{Get script location (using a local copy or not)}
\usage{
get_script_location(local = local_copy, target = "scripts")
}
\arguments{
\item{local}{do you want a local copy on your hard disk to work from}

\item{target}{target path when a local copy is made}
}
\value{
path to script location
}
\description{
Get script location (using a local copy or not)
}
