% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symptom_analysis_data.R
\name{get_symptom_analysis_data}
\alias{get_symptom_analysis_data}
\title{Join the treedata with the symptomdata on the variable MetingKey}
\usage{
get_symptom_analysis_data(trees, symptoms)
}
\arguments{
\item{trees}{dataset containing tree information}

\item{symptoms}{dataset containing symptom information}
}
\value{
data.frame with the joined dataset
}
\description{
Join the treedata with the symptomdata on the variable MetingKey
}
