% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{get_SymptomAnalysisdata}
\alias{get_SymptomAnalysisdata}
\title{Join the treedata eith the symptomdata in==on the variable MetingKey}
\usage{
get_SymptomAnalysisdata(trees, symptoms)
}
\arguments{
\item{trees}{dataset containing tree information}

\item{symptoms}{dataset containing symptom information}
}
\value{
data.frame with the joined dataset
}
\description{
Join the treedata eith the symptomdata in==on the variable MetingKey
}
