% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{get_symptomdata}
\alias{get_symptomdata}
\title{Get the sumptom data}
\usage{
get_symptomdata(channel, jaar, sql, show_query = FALSE)
}
\arguments{
\item{channel}{open DBI connection}

\item{jaar}{the year of which the symptoms are to be read}

\item{sql}{the sql file containing the query for the symptoms}

\item{show_query}{flag if the query should be shown in the standard output}
}
\value{
data.frame with symptom data
}
\description{
Get the sumptom data
}
