% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgmt_functions.R
\name{init_session}
\alias{init_session}
\title{Functie om de instelvariabelen te zetten om de scripts te kunnen runnen}
\usage{
init_session(
  year,
  first_year = 1987,
  first_multiyear = 1995,
  connect_via_db = TRUE,
  outdir = "output",
  fig_width = 7,
  fig_height = 5,
  fig_dpi = 300,
  sen_boot = 200,
  sen_seed = NULL,
  lmer_boot = 200
)
}
\arguments{
\item{year}{laatste datajaar van de analyse}

\item{first_year}{Eerste jaar die meegenomen wordt de heel globale figuren zoals voor de natuurindicatoren}

\item{first_multiyear}{Eerste jaar die meegenomen wordt voor de meerjaarlijkse analyses}

\item{connect_via_db}{haal de data uit de db indien TRUE, anders gebruik lokale files die eerder geïmporteerd werden}

\item{outdir}{output directory relative to working directory}

\item{fig_width}{standaard figuurbreedte in inch}

\item{fig_height}{standaard figuurhoogote in inch}

\item{fig_dpi}{standaard resolutie voor de figuren}

\item{sen_boot}{aantal bootstrap samples om betrouwbaarheidsintervallen op de sen slope te bepalen, indien 0 dan wordt geen bootstrap uitgevoerd}

\item{sen_seed}{chosen seed for the sen calculations, when empty a random seed is chosen}

\item{lmer_boot}{aantal bootstrap samples om  betrouwbaarheidsintervallen voor de lineaire modellen te bepalen, indien 0 dan wordt geen bootstrap uitgevoerd}
}
\value{
maakt verschillende golbale variabelen aan: jaarkeuze, pathkeuze, tweejaarlijks, driejaarlijks, meerjaarlijks, jaren_natuurindicatoren, outdir, connect_via_db, normal_groups, all_groups, extended_groups, groups_multiyear, extra_groups
}
\description{
Functie om de instelvariabelen te zetten om de scripts te kunnen runnen
}
