% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_base_script.R
\name{install_base_script}
\alias{install_base_script}
\title{Krijg het basisscript uit het inbobosvitaliteitspackage in werkdir}
\usage{
install_base_script(output = c("file", "console"))
}
\arguments{
\item{output}{if file copy to 00_base_script.R else show code in console}
}
\value{
script contents in console or base script in a .R file
}
\description{
Krijg het basisscript uit het inbobosvitaliteitspackage in werkdir
}
