% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgmt_functions.R
\name{install_necessary_packages}
\alias{install_necessary_packages}
\title{Installeer benodigde packages}
\usage{
install_necessary_packages()
}
\value{
installed libraries
}
\description{
Installeer benodigde packages
}
