% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_functions.R
\name{pred_sen_slope}
\alias{pred_sen_slope}
\title{Predictions of sen slope}
\usage{
pred_sen_slope(data, sen_boot = 0)
}
\arguments{
\item{data}{data to calculate the sen slope on}

\item{sen_boot}{amount of bootstrap samples for the confidence intervals. If 0 only the estimated values are returned}
}
\value{
data.frame with the sen bootstrap results
}
\description{
Predictions of sen slope
}
