% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{read_species_information}
\alias{read_species_information}
\title{Read species metadata}
\usage{
read_species_information(
  path = "data/tree_reportmetadata.csv",
  local = FALSE,
  load_path = "data/dfSoortInfo.RDS"
)
}
\arguments{
\item{path}{locatie van de csv die de gegevens bevat}

\item{local}{if TRUE read the csv and save the results in load_path else load the data from load_path}

\item{load_path}{path where imported data is saved for load later}
}
\value{
data.frame with species grouping information
}
\description{
Read species metadata
}
