% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summarize_treedata}
\alias{summarize_treedata}
\title{Summarise tree data by predefined groups}
\usage{
summarize_treedata(
  data,
  indiv_id,
  meas_id,
  detail_id = NULL,
  response = NULL,
  count_unique = NULL,
  groups0 = NULL,
  groups1 = NULL,
  groups2 = NULL
)
}
\arguments{
\item{data}{tree information data}

\item{indiv_id}{column name where to find the tree ID}

\item{meas_id}{column name with the measurement ID}

\item{detail_id}{column name with the symptom detail ID}

\item{response}{column name of the response variable}

\item{count_unique}{column name with the unique counts}

\item{groups0}{first grouping variable}

\item{groups1}{second grouping variable}

\item{groups2}{third grouping variable}
}
\value{
dataframe with summarized tree information
}
\description{
Summarise tree data by predefined groups
}
