% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_table.R
\name{wilcox_table}
\alias{wilcox_table}
\title{Wilcoxon test to compare 2 groups from the tree data}
\usage{
wilcox_table(
  data,
  formula = BladverliesNetto ~ LeeftijdsklasseEur,
  alphas = c(0.05, 0.01, 0.001),
  paired = TRUE,
  ...
)
}
\arguments{
\item{data}{data for the calculations}

\item{formula}{the wilcoxon formulat to use}

\item{alphas}{the confidence limits to calculate}

\item{paired}{flag to set the data paired (in conjunction with | in the formula)}

\item{...}{additional parameters}
}
\value{
data.frame with wilcoxon results
}
\description{
Wilcoxon test to compare 2 groups from the tree data
}
