% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symptomdata.R
\name{get_symptomdata}
\alias{get_symptomdata}
\title{Get the symptom data}
\usage{
get_symptomdata(
  channel,
  jaar,
  sqlfile = "package",
  show_query = FALSE,
  local = FALSE,
  load_path = "data/dfSymptoms.RDS"
)
}
\arguments{
\item{channel}{open DBI connection}

\item{jaar}{the year of which the symptoms are to be read}

\item{sqlfile}{the sql file containing the query for the symptoms. When package then us the sql file that is shipped with the package}

\item{show_query}{flag if the query should be shown in the standard output}

\item{local}{whether the data is loaded from load path or imported from the db and then saved in the load_path}

\item{load_path}{path to save/load data}
}
\value{
data.frame with symptom data
}
\description{
Get the symptom data
}
