% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{get_symtomsummary}
\alias{get_symtomsummary}
\title{Find number of symptoms in the whole dataset}
\usage{
get_symtomsummary(trees, symptoms)
}
\arguments{
\item{trees}{dataset with the tree information}

\item{symptoms}{dataset with the symptom information}
}
\value{
data.frame containng the number of symptoms per tree
}
\description{
Find number of symptoms in the whole dataset
}
