% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{get_treedata}
\alias{get_treedata}
\title{Get the basic tree information from the db}
\usage{
get_treedata(
  channel,
  jaar,
  tree_indeling,
  sql,
  show_query = TRUE,
  tweejaarlijks = NULL,
  driejaarlijks = NULL
)
}
\arguments{
\item{channel}{open DBI SQL server channel}

\item{jaar}{current year}

\item{tree_indeling}{file containing the tree metadata}

\item{sql}{file containing the sql query}

\item{show_query}{flag if query should be shown in standard output}

\item{tweejaarlijks}{optional the year to compare the current year against}

\item{driejaarlijks}{optional the 2 years to compare the current year against}
}
\value{
a data.frame containing all the tree information from the query
}
\description{
Get the basic tree information from the db
}
