% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_treedata.R
\name{get_treedata}
\alias{get_treedata}
\title{Get the basic tree information from the db}
\usage{
get_treedata(
  channel,
  jaar,
  tree_indeling,
  sqlfile = "package",
  load_path = "data/dfTrees.RDS",
  show_query = FALSE,
  local = FALSE
)
}
\arguments{
\item{channel}{open DBI SQL server channel}

\item{jaar}{current year}

\item{tree_indeling}{file containing the tree metadata}

\item{sqlfile}{file containing the sql query}

\item{load_path}{path to save/load data to/from}

\item{show_query}{flag if query should be shown in standard output}

\item{local}{if TRUE get the data from load_path else get the data from the DB and save a copy in load_path}
}
\value{
a data.frame containing all the tree information from the query
}
\description{
Get the basic tree information from the db
}
